bug detectors: split up tests into several files
