+++ tests/bug-detectors/general.test.js
@@ -15,16 +15,16 @@
  */
 
 /* eslint no-undef: 0 */
-// eslint-disable-next-line @typescript-eslint/no-var-requires
-const { FuzzTestBuilder } = require("./helpers.js");
+const {
+	FuzzTestBuilder,
+	FuzzingExitCode,
+	JestRegressionExitCode,
+	// eslint-disable-next-line @typescript-eslint/no-var-requires
+} = require("./helpers.js");
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const path = require("path");
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const fs = require("fs");
-// This is used to distinguish an error thrown during fuzzing from other errors, such as wrong
-// `fuzzEntryPoint` (which would return a "1")
-const FuzzingExitCode = "77";
-const JestRegressionExitCode = "1";
 
 describe("General tests", () => {
 	const bugDetectorDirectory = path.join(__dirname, "general");
@@ -358,303 +358,5 @@ describe("General tests", () => {
 			.build();
 		fuzzTest.execute();
 		expect(fs.existsSync(friendlyFilePath)).toBeTruthy();
-	});
-});
-
-describe("Command injection", () => {
-	const bugDetectorDirectory = path.join(__dirname, "command-injection");
-	const friendlyFilePath = path.join(bugDetectorDirectory, "FRIENDLY");
-
-	// Delete files created by the tests.
-	beforeEach(() => {
-		fs.rmSync(friendlyFilePath, { force: true });
-	});
-
-	it("exec with EVIL command", () => {
-		const fuzzTest = new FuzzTestBuilder()
-			.sync(false)
-			.fuzzEntryPoint("execEVIL")
-			.dir(bugDetectorDirectory)
-			.build();
-		expect(() => {
-			fuzzTest.execute();
-		}).toThrow(FuzzingExitCode);
-		expect(fs.existsSync(friendlyFilePath)).toBeFalsy();
-	});
-
-	it("exec with FRIENDLY command", () => {
-		const fuzzTest = new FuzzTestBuilder()
-			.sync(false)
-			.fuzzEntryPoint("execFRIENDLY")
-			.dir(bugDetectorDirectory)
-			.build();
-		fuzzTest.execute();
-		expect(fs.existsSync(friendlyFilePath)).toBeTruthy();
-	});
-
-	it("execFile with EVIL file", () => {
-		const fuzzTest = new FuzzTestBuilder()
-			.sync(false)
-			.fuzzEntryPoint("execFileEVIL")
-			.dir(bugDetectorDirectory)
-			.build();
-		expect(() => {
-			fuzzTest.execute();
-		}).toThrow(FuzzingExitCode);
-		expect(fs.existsSync(friendlyFilePath)).toBeFalsy();
-	});
-
-	it("execFile with FRIENDLY file", () => {
-		const fuzzTest = new FuzzTestBuilder()
-			.sync(false)
-			.fuzzEntryPoint("execFileFRIENDLY")
-			.dir(bugDetectorDirectory)
-			.build();
-		fuzzTest.execute();
-		expect(fs.existsSync(friendlyFilePath)).toBeTruthy();
-	});
-
-	it("execFileSync with EVIL file", () => {
-		const fuzzTest = new FuzzTestBuilder()
-			.sync(false)
-			.fuzzEntryPoint("execFileSyncEVIL")
-			.dir(bugDetectorDirectory)
-			.build();
-		expect(() => {
-			fuzzTest.execute();
-		}).toThrow(FuzzingExitCode);
-		expect(fs.existsSync(friendlyFilePath)).toBeFalsy();
-	});
-
-	it("execFileSync with FRIENDLY file", () => {
-		const fuzzTest = new FuzzTestBuilder()
-			.sync(false)
-			.fuzzEntryPoint("execFileSyncFRIENDLY")
-			.dir(bugDetectorDirectory)
-			.build();
-		fuzzTest.execute();
-		expect(fs.existsSync(friendlyFilePath)).toBeTruthy();
-	});
-
-	it("spawn with EVIL command", () => {
-		const fuzzTest = new FuzzTestBuilder()
-			.sync(false)
-			.fuzzEntryPoint("spawnEVIL")
-			.dir(bugDetectorDirectory)
-			.build();
-		expect(() => {
-			fuzzTest.execute();
-		}).toThrow(FuzzingExitCode);
-		expect(fs.existsSync(friendlyFilePath)).toBeFalsy();
-	});
-
-	it("spawn with FRIENDLY command", () => {
-		const fuzzTest = new FuzzTestBuilder()
-			.sync(false)
-			.fuzzEntryPoint("spawnFRIENDLY")
-			.dir(bugDetectorDirectory)
-			.build();
-		fuzzTest.execute();
-		expect(fs.existsSync(friendlyFilePath)).toBeTruthy();
-	});
-
-	it("spawnSync with EVIL command", () => {
-		const fuzzTest = new FuzzTestBuilder()
-			.sync(false)
-			.fuzzEntryPoint("spawnSyncEVIL")
-			.dir(bugDetectorDirectory)
-			.build();
-		expect(() => {
-			fuzzTest.execute();
-		}).toThrow(FuzzingExitCode);
-		expect(fs.existsSync(friendlyFilePath)).toBeFalsy();
-	});
-
-	it("spawnSync with FRIENDLY command", () => {
-		const fuzzTest = new FuzzTestBuilder()
-			.sync(false)
-			.fuzzEntryPoint("spawnSyncFRIENDLY")
-			.dir(bugDetectorDirectory)
-			.build();
-		fuzzTest.execute();
-		expect(fs.existsSync(friendlyFilePath)).toBeTruthy();
-	});
-
-	it("fork with EVIL command", () => {
-		const fuzzTest = new FuzzTestBuilder()
-			.sync(false)
-			.fuzzEntryPoint("forkEVIL")
-			.dir(bugDetectorDirectory)
-			.build();
-		expect(() => {
-			fuzzTest.execute();
-		}).toThrow(FuzzingExitCode);
-		expect(fs.existsSync(friendlyFilePath)).toBeFalsy();
-	});
-
-	it("fork with FRIENDLY command", () => {
-		const fuzzTest = new FuzzTestBuilder()
-			.sync(false)
-			.fuzzEntryPoint("forkFRIENDLY")
-			.dir(bugDetectorDirectory)
-			.build();
-		fuzzTest.execute();
-		expect(fs.existsSync(friendlyFilePath)).toBeTruthy();
-	});
-});
-
-describe("Path Traversal", () => {
-	const SAFE = "../safe_path/";
-	const EVIL = "../evil_path/";
-
-	beforeEach(() => {
-		fs.rmSync(SAFE, { recursive: true, force: true });
-	});
-
-	const bugDetectorDirectory = path.join(__dirname, "path-traversal");
-
-	it("openSync with EVIL path", () => {
-		const fuzzTest = new FuzzTestBuilder()
-			.sync(true)
-			.fuzzEntryPoint("PathTraversalFsOpenEvilSync")
-			.dir(bugDetectorDirectory)
-			.build();
-		expect(() => {
-			fuzzTest.execute();
-		}).toThrow(FuzzingExitCode);
-	});
-
-	it("open with EVIL path", () => {
-		const fuzzTest = new FuzzTestBuilder()
-			.sync(false)
-			.fuzzEntryPoint("PathTraversalFsOpenEvilAsync")
-			.dir(bugDetectorDirectory)
-			.build();
-		expect(() => {
-			fuzzTest.execute();
-		}).toThrow(FuzzingExitCode);
-	});
-
-	it("mkdirSync with EVIL path", () => {
-		const fuzzTest = new FuzzTestBuilder()
-			.sync(true)
-			.fuzzEntryPoint("PathTraversalFsMkdirEvilSync")
-			.dir(bugDetectorDirectory)
-			.build();
-		expect(() => {
-			fuzzTest.execute();
-		}).toThrow(FuzzingExitCode);
-		expect(fs.existsSync(EVIL)).toBeFalsy();
-		expect(fs.existsSync(SAFE)).toBeFalsy();
-	});
-
-	it("mkdirSync with SAFE path", () => {
-		const fuzzTest = new FuzzTestBuilder()
-			.sync(true)
-			.fuzzEntryPoint("PathTraversalFsMkdirSafeSync")
-			.dir(bugDetectorDirectory)
-			.build();
-		fuzzTest.execute();
-		expect(fs.existsSync(SAFE)).toBeTruthy();
-		expect(fs.existsSync(EVIL)).toBeFalsy();
-	});
-
-	it("mkdirAsync with EVIL path", () => {
-		const fuzzTest = new FuzzTestBuilder()
-			.sync(false)
-			.fuzzEntryPoint("PathTraversalFsMkdirEvilAsync")
-			.dir(bugDetectorDirectory)
-			.build();
-		expect(() => {
-			fuzzTest.execute();
-		}).toThrow(FuzzingExitCode);
-		expect(fs.existsSync(EVIL)).toBeFalsy();
-		expect(fs.existsSync(SAFE)).toBeFalsy();
-	});
-
-	it("mkdirAsync with SAFE path", () => {
-		const fuzzTest = new FuzzTestBuilder()
-			.sync(false)
-			.fuzzEntryPoint("PathTraversalFsMkdirSafeAsync")
-			.dir(bugDetectorDirectory)
-			.build();
-		fuzzTest.execute();
-		expect(fs.existsSync(SAFE)).toBeTruthy();
-		expect(fs.existsSync(EVIL)).toBeFalsy();
-	});
-
-	it("mkdir PROMISES with SAFE path", () => {
-		const fuzzTest = new FuzzTestBuilder()
-			.sync(false)
-			.fuzzEntryPoint("PathTraversalFspMkdirSafeAsync")
-			.dir(bugDetectorDirectory)
-			.build();
-		fuzzTest.execute();
-		expect(fs.existsSync(SAFE)).toBeTruthy();
-		expect(fs.existsSync(EVIL)).toBeFalsy();
-	});
-
-	it("mkdir PROMISES with EVIL path", () => {
-		const fuzzTest = new FuzzTestBuilder()
-			.sync(false)
-			.fuzzEntryPoint("PathTraversalFspMkdirEvilAsync")
-			.dir(bugDetectorDirectory)
-			.build();
-		expect(() => {
-			fuzzTest.execute();
-		}).toThrow(FuzzingExitCode);
-		expect(fs.existsSync(EVIL)).toBeFalsy();
-		expect(fs.existsSync(SAFE)).toBeFalsy();
-	});
-
-	it("open PROMISES with EVIL path", () => {
-		const fuzzTest = new FuzzTestBuilder()
-			.sync(false)
-			.fuzzEntryPoint("PathTraversalFspOpenEvilAsync")
-			.dir(bugDetectorDirectory)
-			.build();
-		expect(() => {
-			fuzzTest.execute();
-		}).toThrow(FuzzingExitCode);
-	});
-
-	it("joinSync with EVIL path", () => {
-		const fuzzTest = new FuzzTestBuilder()
-			.sync(true)
-			.fuzzEntryPoint("PathTraversalJoinEvilSync")
-			.dir(bugDetectorDirectory)
-			.build();
-		expect(() => {
-			fuzzTest.execute();
-		}).toThrow(FuzzingExitCode);
-	});
-
-	it("joinSync with SAFE path", () => {
-		const fuzzTest = new FuzzTestBuilder()
-			.sync(true)
-			.fuzzEntryPoint("PathTraversalJoinSafeSync")
-			.dir(bugDetectorDirectory)
-			.build();
-		fuzzTest.execute();
-	});
-
-	it("join with EVIL path", () => {
-		const fuzzTest = new FuzzTestBuilder()
-			.sync(false)
-			.fuzzEntryPoint("PathTraversalJoinEvilAsync")
-			.dir(bugDetectorDirectory)
-			.build();
-		expect(() => {
-			fuzzTest.execute();
-		}).toThrow(FuzzingExitCode);
-	});
-
-	it("join with SAFE path", () => {
-		const fuzzTest = new FuzzTestBuilder()
-			.sync(false)
-			.fuzzEntryPoint("PathTraversalJoinSafeAsync")
-			.dir(bugDetectorDirectory)
-			.build();
-		fuzzTest.execute();
 	});
 });
